chrome.manifest: fix ambiguous path to modules:
This is to fix problem which in some cases makes https-finder not
functional: it doesn't probe for HTTPS and show no prompts or errors
although its preferences are accessible.

As far as I'm aware this is safe change and other extensions often
define path to modules without prepending any paths.